Moved stuff to util file.
